pyproject.toml
[tool.poetry]
name = "cw-7-drf"
version = "0.1.0"
description = ""
authors = ["StepanShurmin <shurmin.sa@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^4.2.8"
djangorestframework = "^3.14.0"
pillow = "^10.1.0"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.0"
redis = "^5.0.1"
celery = "^5.3.6"
django-celery-beat = "^2.5.0"
django-cors-headers = "^4.3.1"
setuptools = "^69.0.2"
black = {extras = ["d"], version = "^23.11.0"}
django-filter = "^23.5"
djangorestframework-simplejwt = "^5.3.1"
requests = "^2.31.0"
drf-yasg = "^1.21.7"
coverage = "^7.3.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
